pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hydra-repo-syncer"
version = "0.4.0"
description = "Provision a repo across self-hosted GitLab, GitLab.com, and GitHub with push mirroring."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "aanand" }]
dependencies = [
    "requests>=2.31",
    "urllib3>=1.26",
    "typer>=0.12",
    "pyyaml>=6.0",
    "keyring>=24.0",
    "python-dotenv>=1.0",
    "questionary>=2.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-xdist>=3.5",
    "requests-mock>=1.11",
    "ruff>=0.6",
]

[project.scripts]
hydra = "hydra.cli:app"

[tool.setuptools.packages.find]
include = ["hydra*"]
exclude = ["tests*", "venv*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py39"
extend-exclude = ["venv", ".venv", "build", "dist"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
# Project convention: prefer `typing` module forms (Dict, List, Optional, Union)
# over PEP 585 / PEP 604 lowercase generics. Disable the auto-modernization rules.
ignore = ["E501", "UP006", "UP007", "UP035", "UP045"]

[tool.ruff.lint.per-file-ignores]
# Typer requires `typer.Option(...)` as parameter defaults (B008) and
# evaluates annotations at runtime via get_type_hints, which can't resolve
# `X | None` on Python 3.9 (UP007/UP045). Keep CLI command modules compatible.
"hydra/cli/*.py" = ["B008", "UP007", "UP045"]

[tool.coverage.run]
source = ["hydra"]
omit = ["hydra/__main__.py"]

[tool.coverage.report]
fail_under = 30
show_missing = true
